Split the operator-precedence module into many more files
