+++ Sources/SwiftOperatorPrecedence/Operator.swift
@@ -1,0 +1,50 @@
+//===------------------ Operator.swift ------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+import SwiftSyntax
+
+/// Names an operator.
+///
+/// TODO: For now, we'll use strings, but we likely want to move this to a
+/// general notion of an Identifier.
+public typealias OperatorName = String
+
+/// Describes the kind of an operator.
+public enum OperatorKind: String {
+  /// Infix operator such as the + in a + b.
+  case infix
+
+  /// Prefix operator such as the - in -x.
+  case prefix
+
+  /// Postfix operator such as the ! in x!.
+  case postfix
+}
+
+/// Describes an operator.
+public struct Operator {
+  public let kind: OperatorKind
+  public let name: OperatorName
+  public let precedenceGroup: PrecedenceGroupName?
+  public let syntax: OperatorDeclSyntax?
+
+  public init(
+    kind: OperatorKind, name: OperatorName,
+    precedenceGroup: PrecedenceGroupName?,
+    syntax: OperatorDeclSyntax? = nil
+  ) {
+    self.kind = kind
+    self.name = name
+    self.precedenceGroup = precedenceGroup
+    self.syntax = syntax
+  }
+}
