+++ Sources/SwiftOperatorPrecedence/OperatorPrecedence+Semantics.swift
@@ -1,4 +1,4 @@
-//===-------------- OperatorPrecedenceParsing.swift -----------------------===//
+//===-------------- OperatorPrecedence+Semantics.swift --------------------===//
 //
 // This source file is part of the Swift.org open source project
 //
